refactor: extract canvas rendering hook
